requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
WTForms==3.1.2
email-validator==2.1.0.post1
SQLAlchemy==2.0.23
python-dotenv==1.0.1
reportlab==4.0.7
gunicorn==21.2.0
